requirements.txt
qdrant-client
langchain-huggingface
langchain
pymupdf
numpy
regex
pyvi
sentence-transformers
torchvision
torchaudio
python-docx
fastapi
python-multipart
uvicorn
docx2pdf